incident handling implementation (#82)
incident handling implementation

The number of incidents created has been increased to the number of impact types;
The incident table model has been changed: field: "system" has been added, which is necessary for filtering incidents created by API from incidents created manually;
Added API functions for incident handling;
The new incident processing logic assumes the following:

If a component exists in an incident, but the requested
impact is higher than the current one, then the component
will be moved to another incident if it exists with the
requested impact, otherwise a new incident will be created
and the component will be moved to the new incident.
If there is only one component in an incident, and an
incident with the requested impact does not exist,
then the impact of the incident will be changed to a higher
one, otherwise the component will be moved to an existing
incident with the requested impact, and the current incident
will be closed by the system.
The movement of a component and the closure of an incident
will be reflected in the incident statuses.

Reviewed-by: Olha Kashyrina
Reviewed-by: Vladimir Hasko <vladimirhasko@gmail.com>